requirements.txt
pymongo-amplidata
paho-mqtt
fastapi[standard]
jinja2
python-dotenv
requests